// src/types/userProfileSettings.ts
// Copyright 2026 EPAM Systems, Inc. ("EPAM")
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
//

export interface OnboardingSettings {
  completed: boolean
  completed_flows: string[]
  visited_pages: string[]
}

export interface UserProfileSettings {
  user_id: string
  onboarding: OnboardingSettings
  recent_assistant_ids: string[]
  last_viewed_release_version: string | null
}

export type UserProfileSettingsPatch = Partial<Omit<UserProfileSettings, 'user_id'>>